+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Removed
+
+- Removed `!command` CTA to use slash commands and legacy prefix logic.
+
 ## [v7.5.2](https://github.com/lexicalunit/spellbot/releases/tag/v7.5.2) - 2021-11-23
 
 ### Changed